+++ packages/doctor/test/android-tools-info.ts
@@ -56,6 +56,7 @@ describe("androidToolsInfo", () => {
 						"31.0.0",
 						"32.0.0",
 						"33.0.0",
+						"34.0.0",
 					];
 				} else {
 					return [
@@ -67,6 +68,7 @@ describe("androidToolsInfo", () => {
 						"android-31",
 						"android-32",
 						"android-33",
+						"android-34",
 					];
 				}
 			},
@@ -91,18 +93,18 @@ describe("androidToolsInfo", () => {
 			assert.equal(toolsInfo.compileSdkVersion, 30);
 		});
 
-		it("runtime 8.1.1 - 30", () => {
+		it("runtime < 8.2.0 - 30", () => {
 			const androidToolsInfo = getAndroidToolsInfo("8.1.1");
 			const toolsInfo = androidToolsInfo.getToolsInfo({ projectDir: "test" });
 
 			assert.equal(toolsInfo.compileSdkVersion, 30);
 		});
 
-		it("runtime 8.2.0 - 32", () => {
+		it("runtime >8.2.0 - latest", () => {
 			const androidToolsInfo = getAndroidToolsInfo("8.2.0");
 			const toolsInfo = androidToolsInfo.getToolsInfo({ projectDir: "test" });
 
-			assert.equal(toolsInfo.compileSdkVersion, 33);
+			assert.equal(toolsInfo.compileSdkVersion, 34);
 		});
 	});
 
@@ -134,9 +136,9 @@ describe("androidToolsInfo", () => {
 			assertSupportedRange("8.1.0", min, max);
 		});
 
-		it("runtime 8.2.0 should support android-17 - android-33", () => {
+		it("runtime 8.2.0 should support android-17 - android-34", () => {
 			const min = 17;
-			const max = 33;
+			const max = 34;
 			assertSupportedRange("8.2.0", min, max);
 			assertSupportedRange("8.3.0", min, max);
 		});